SFTENT: don't abort enter just because a key up event arrives shortly after the key is pressed
I noticed that when typing clear that sometimes the ENTER after the command (as KC_SFTENT) got lost.  I'm not sure why the code does this, but when any event arrives that's not SFTENT, the SFTENT is immediately cancelled and we can never get to the state where we emit ENTER.  In this case, it was the R up event cancelling it, which seems pointless to me.

A better fix might be to check the timer and allow key up events for a short period of time after arming the enter key, but after that timer expires, to cancel.  I'm not sure what would cause someone to need this, so I didn't do it.  Obviously I will accidentally press enter somewhere inappropirate at some point and then implement this feature.